requirements.txt
blinker==1.6.2
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
httplib2==0.22.0
humanize==4.8.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
oauthlib==3.2.2
packaging==23.1
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.7
